+++ shardis-ui/src/main/frontend/platform/browser/providers.ts
@@ -6,7 +6,7 @@ import {FORM_PROVIDERS} from 'angular2/common';
 // Angular 2 Http
 import {HTTP_PROVIDERS} from 'angular2/http';
 // Angular 2 Router
-import {ROUTER_PROVIDERS, LocationStrategy, HashLocationStrategy} from 'angular2/router';
+import {ROUTER_PROVIDERS, LocationStrategy, PathLocationStrategy} from 'angular2/router';
 
 // Angular 2 Material
 // import {MdRadioDispatcher} from '@angular2-material/radio/radio_dispatcher';
@@ -23,7 +23,7 @@ export const APPLICATION_PROVIDERS = [
   ...HTTP_PROVIDERS,
   // ...MATERIAL_PROVIDERS,
   ...ROUTER_PROVIDERS,
-  provide(LocationStrategy, { useClass: HashLocationStrategy }),
+  provide(LocationStrategy, { useClass: PathLocationStrategy }),
 ];
 
 export const PROVIDERS = [